A few reworks and adjustments
I reworked the buddyIDs array to store only player UIDs instead of IDs. I also reworked buddy deleting so that it deletes buddies from the array.

+++ src/Player.hpp
@@ -69,5 +69,5 @@ struct Player {
     int32_t iGroupConditionBitFlag;
 
     int buddyCnt;
-    int32_t buddyIDs[50];
+    int64_t buddyIDs[50];
 };
